updated hibernate deployed

+++ src/main/java/com/revature/bibimbop/credit_card/CreditCardDTO.java
@@ -9,7 +9,7 @@ public class CreditCardDTO {
     private Integer cvv;
     private String expDate;
     private Integer zip;
-    private double limits;
+    private BigDecimal limits;
     private String customerUsername;
     private String tableSelection;
     private String newCellName;
@@ -59,7 +59,7 @@ public class CreditCardDTO {
         return limits;
     }
 
-    public void setLimits(Double limits) {
+    public void setLimits(BigDecimal limits) {
         this.limits = limits;
     }
